Rename init-stich to init-stitch
